refactor: added models
